#896 moved the Context Element Adapter to Sgraph plugin. Context Element Adapter removed from XtextStyledEditor and XTextDirectEditmanager, it is set from the calling EditPart.
