+++ plugins/org.yakindu.base.xtext.utils.gmf/src/org/yakindu/base/xtext/utils/gmf/viewers/XtextStyledTextCellEditorEx.java
@@ -12,7 +12,6 @@
 package org.yakindu.base.xtext.utils.gmf.viewers;
 
 import org.yakindu.base.xtext.utils.jface.viewers.XtextStyledTextCellEditor;
-import org.yakindu.base.xtext.utils.jface.viewers.ContextElementAdapter.IContextElementProvider;
 import org.yakindu.base.xtext.utils.jface.viewers.context.IXtextFakeContextResourcesProvider;
 
 import com.google.inject.Injector;
@@ -36,11 +35,6 @@ public class XtextStyledTextCellEditorEx extends XtextStyledTextCellEditor {
 
 	public XtextStyledTextCellEditorEx(int style, Injector injector) {
 		super(style, injector);
-	}
-
-	public XtextStyledTextCellEditorEx(int style, Injector injector,
-			IContextElementProvider provider) {
-		super(style, injector, provider);
 	}
 
 	/**
