+++ test-plugins/org.yakindu.sct.model.stext.resource.test/META-INF/MANIFEST.MF
@@ -6,6 +6,5 @@ Bundle-Version: 2.7.1.qualifier
 Fragment-Host: org.yakindu.sct.model.resource
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Require-Bundle: org.junit,
- org.yakindu.base.xtext.utils.jface,
  org.yakindu.sct.test.models
 Bundle-Vendor: statecharts.org
